feature/37 refactor: file name MessageItem -> Message
