+++ packages/ui-kit/Message/index.ts
@@ -1,0 +1,2 @@
+export { default as Message } from './Message';
+export * from './types';
