+++ packages/ui-kit/Message/types.ts
@@ -1,4 +1,4 @@
-export enum MessageItemVariant {
+export enum MessageVariant {
     PRIMARY = 'primary',
     SECONDARY = 'secondary',
 }
